Init
We are experimenting with a migration of all of hackint towards a common
configuration management via NixOS.

The plan is to create a staging network with the components that will
operate our network in the near future.

Includes a module to run solanum that relies on a boilerplate network
module and a wireguard module, that will link nodes between each other,
that is yet to come.

Co-Authored-By: Dustin Frisch <fooker@lab.sh>